+++ packages/gauges/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pancakeswap/gauges
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [205c08713]
+  - @pancakeswap/tokens@0.6.0
+  - @pancakeswap/v3-sdk@3.7.5
+
 ## 0.0.4
 
 ### Patch Changes